+++ Command/UpdateCorrectAddressDataCommand.php
@@ -11,60 +11,68 @@ namespace MauticPlugin\MauticEnhancerBundle\Command;
 use League\Flysystem\Filesystem;
 use League\Flysystem\Sftp\SftpAdapter;
 use Mautic\CoreBundle\Command\ModeratedCommand;
-use MauticPlugin\MauticEnhancerBundle\Integration\ExpirianCorrectAddressIntegration as ECAI;
+use MauticPlugin\MauticEnhancerBundle\Helper\EnhancerHelper;
+use MauticPlugin\MauticEnhancerBundle\Integration\CorrectAddressIntegration as CAI;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class UpdateCorrectAddressData extends ModeratedCommand
+class UpdateCorrectAddressDataCommand extends ModeratedCommand
 {
     protected function configure()
     {
-        $this
-            ->setName('mautic:integration:enhancer:updatecorrectaddress')
-            ->setDescription('Installs the latest data files available from Expirian')
-            ->setHelp('This command will download and replace the data files used by CorrectAddress. These are proprietary files available from Expirian');
+        $this->setName('mautic:integration:enhancer:updatecorrectaddress');
+        $this->setDescription('Installs the latest data files available from Expirian');
+        $this->setHelp('This command will download and replace the data files used by CorrectAddress. These are proprietary files available from Expirian');
     }
 
     /**
      * @param InputInterface  $input
      * @param OutputInterface $output
      *
-     * @return bool
+     * @return int|null|void
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         try {
-            /** @var \MauticPlugin\MauticEnhancerBundle\Integration\ExpirianCorrectAddressIntegration $correctAddress */
-            $correctAddress = $this->getContainer()
-                ->get('mautic.enhancer.helper.enhancer')
-                ->getIntegration('mautic.enhancer.integration.correctaddress');
+            echo 'Starting Expirian data update';
 
-            $settings = $correctAddress->getSupportedFeatures();
-            $keys     = $correctAddress->getKeys();
+            /** @var \MauticPlugin\MauticEnhancerBundle\Helper\EnhancerHelper $correctAddress */
+            $enhancerHelper = new EnhancerHelper($this->getContainer()->get('mautic.helper.integration'));
+            $correctAddress = $enhancerHelper->getIntegration('CorrectAddress');
+            $settings       = $correctAddress->getIntegrationSettings()->getFeatureSettings();
+            $keys           = $correctAddress->getKeys();
 
-            $SftpAdapter = new SftpAdapter([
-                'host'            => $settings[ECAI::CA_REMOTE_HOST],
-                'port'            => $settings[ECAI::CA_REMOTE_PORT],
-                'hostFingerprint' => $keys[ECAI::CA_REMOTE_FNGR],
-                'username'        => $keys[ECAI::CA_REMOTE_USER],
-                'password'        => $keys[ECAI::CA_REMOTE_PSWD],
-                'root'            => $settings[ECAI::CA_REMOTE_PATH],
-            ]);
-            $client = new Filesystem($SftpAdapter);
+            $client = new Filesystem(new SftpAdapter([
+                'host'            => $settings[CAI::CA_REMOTE_HOST],
+                'port'            => $settings[CAI::CA_REMOTE_PORT],
+                'root'            => $settings[CAI::CA_REMOTE_PATH],
+                'username'        => $keys[CAI::CA_REMOTE_USER],
+                'password'        => $keys[CAI::CA_REMOTE_PSWD],
+                'hostFingerprint' => $keys[CAI::CA_REMOTE_FNGR],
+            ]));
+            echo 'Created SFTP client'.PHP_EOL;
 
             //copy the remote archive locally
             $tempfile = tempnam(sys_get_temp_dir(), 'ca_');
-            $client->copy($settings[ECAI::CA_REMOTE_FILE], $tempfile);
+            $client->copy($settings[CAI::CA_REMOTE_FILE], $tempfile);
+            echo 'Copied data archive to '.$tempfile.' on local filesystem'.PHP_EOL;
 
             //remove the old files
-            $this->cleanDir($settings[ECAI::CA_CORRECTA_DATA]);
+            $this->cleanDir($settings[CAI::CA_CORRECTA_DATA]);
+            echo $settings[CAI::CA_CORRECTA_DATA].' is ready for writing'.PHP_EOL;
 
             $extractor = new \ZipArchive();
             $extractor->open($tempfile);
-            $extractor->extractTo($settings[ECAI::CA_CORRECTA_DATA]);
+            $extractor->extractTo($settings[CAI::CA_CORRECTA_DATA]);
             $extractor->close();
             unlink($tempfile);
+            echo 'Expirian data update complete'.PHP_EOL;
+
+            return 0;
         } catch (\Exception $e) {
+            echo $e->getMessage();
+
+            return $e->getCode();
         }
     }
 
@@ -83,11 +91,10 @@ class UpdateCorrectAddressData extends ModeratedCommand
             return mkdir($dirName, 0755, true);
         }
 
-        $root     = new \RecursiveDirectoryIterator($dirName, \RecursiveDirectoryIterator::SKIP_DOTS);
-        $contents = new \RecursiveIteratorIterator($root, \RecursiveIteratorIterator::CHILD_FIRST);
+        $root = new \RecursiveDirectoryIterator($dirName, \RecursiveDirectoryIterator::SKIP_DOTS);
+        $ls   = new \RecursiveIteratorIterator($root, \RecursiveIteratorIterator::CHILD_FIRST);
 
-        /** @var \SplFileInfo $file */
-        foreach ($contents as $file) {
+        foreach ($ls as $file) {
             $file->isDir() ? rmdir($file->getRealPath()) : unlink($file->getRealPath());
         }
