+++ Integration/RandomIntegration.php
@@ -12,6 +12,8 @@
 namespace MauticPlugin\MauticEnhancerBundle\Integration;
 
 use Mautic\LeadBundle\Entity\Lead;
+use Symfony\Component\Form\Extension\Core\Type\HiddenType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 
 /**
  * Class RandomIntegration.
@@ -52,7 +54,7 @@ class RandomIntegration extends AbstractEnhancerIntegration
         if ('features' === $formArea && !isset($data['random_field_name'])) {
             $builder->add(
                 'random_field_name',
-                \Symfony\Component\Form\Extension\Core\Type\TextType::class,
+                TextType::class,
                 [
                     'label' => $this->translator->trans('mautic.plugin.random.field_name.label'),
                     'attr'  => [
@@ -63,7 +65,7 @@ class RandomIntegration extends AbstractEnhancerIntegration
             )
             ->add(
                 'autorun_enabled',
-                \Symfony\Component\Form\Extension\Core\Type\HiddenType::class,
+                HiddenType::class,
                 [
                     'data' => true,
                 ]
