+++ Translations/en_US/messages.ini
@@ -47,13 +47,14 @@ mautic.enhancer.integration.genderfromname.failure = "Integration failed to Crea
 mautic.enhancer.integration.genderfromname.custom_note = "<p>In order for this Enhancer to work properly, run <blockquote><strong>app/console mautic:integration:enhancer:installgendernames</strong></blockquote></p> This may also resolve issues creating the reference table."
 
 ;CorrectAddress
-mautic.integration.correctaddress.data_server = "Data Host"
-mautic.integration.correctaddress.data_port   = "Port"
-mautic.integration.correctaddress.data_path   = "Path to data file"
-mautic.integration.correctaddress.data_file   = "Data File"
-mautic.integration.correctaddress.correcta_path = "Path to Processor"
-mautic.integration.correctaddress.correcta_cmd  = "Processor"
-mautic.integration.correctaddress.correcta_data = "Path to Data files"
-mautic.integration.correctaddress.username  = "Host User"
-mautic.integration.correctaddress.password  = "Host Password"
-mautic.integration.correctaddress.fingerprint = "Host Fingerprimt"
+mautic.enhancer.correctaddress.username      = "Host User"
+mautic.enhancer.correctaddress.password      = "Host Password"
+mautic.enhancer.correctaddress.fingerprint   = "Host Fingerprimt"
+
+mautic.enhancer.correctaddress.data_server   = "Remote Host"
+mautic.enhancer.correctaddress.data_port     = "Remote Port"
+mautic.enhancer.correctaddress.data_path     = "Remote Path"
+mautic.enhancer.correctaddress.data_file     = "Remote Archive"
+
+mautic.enhancer.correctaddress.correcta_cmd  = "Executable Path"
+mautic.enhancer.correctaddress.correcta_data = "Local Data Path"
